requirements.txt
google-play-scraper
groq
google-genai
requests
python-dotenv
